refactor: move functions to util
